Cleanup collision warnings.
